requirements.txt
Flask==2.3.0
opencv-python==4.6.0.66
numpy==1.23.0
face_recognition==1.3.0
Pillow==9.2.0
requests==2.28.1
flask-cors==3.1.1
mediapipe==0.9.0
gunicorn==20.1.0